Provide CLI options to control the verbosity of log output.
If `--verbose` is specified: allow TRACE and above (i.e., everything).
If `--quiet`   is specified: allow ERROR and above.
If `--silent`  is specified: allow NOTHING.

If none of the above options are specified, allow INFO and above.

If more than one of the above options is specified, display an error and quit.